+++ svelte/src/lib/Animation/ConfigItems/Numerical.svelte
